+++ src/components/Pages/Info/index.jsx
@@ -1,8 +1,9 @@
 import { useContext } from "react"
 import { GameContext } from "../../../context/GameContext"
+import { DataContext } from "../../../context/DataContext"
 
 export default function (){
-    const {bgModal} = useContext(GameContext)
+    const {bgModal} = useContext(DataContext)
     return (
         <div className={`w-[80%] h-fit pt-2 pb-16 px-4 lg:w-[55rem] lg:items-start items-center flex flex-col rounded-3xl lg:rounded-xl flex-wrap gap-8 ${bgModal}`}>
             <h1 className="text-xl font-medium">Infinity Devil Tower - Version 1.0.0</h1>